requirements.txt
appdirs==1.4.3
dill==0.2.6
Django==1.10.6
django-generic-positions==0.2.2
django-hvad==1.7.0
django-libs==1.67.12
django-review==1.9.6
django-user-media==1.2.4
easy-thumbnails==2.3
enum34==1.1.6
future==0.16.0
gapic-google-cloud-vision-v1==0.90.3
google-api-python-client==1.6.2
google-auth==0.9.0
google-auth-httplib2==0.0.2
google-cloud-core==0.23.1
google-cloud-vision==0.23.3
google-gax==0.15.8
googleapis-common-protos==1.5.2
grpcio==1.2.0
httplib2==0.10.3
oauth2client==3.0.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
ply==3.8
proto-google-cloud-vision-v1==0.90.3
protobuf==3.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
pyreadline==2.1
rsa==3.4.2
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
